requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
py-clob-client>=0.17.0
web3>=6.20.0
eth-account>=0.10.0
